services/api_gateway/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
pandas
requests